+++ CHANGELOG.md
@@ -4,6 +4,25 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+------------------------------------------------------------------------------
+
+## [0.0.2] - 2023-11-08
+
+Release of the library according to the new publication []().
+
+### Added
+
+- :computer: [Python] Geometry submodule for the computation of the shape area (ref. `deepskin/geometry.py`)
+- :closed_book: [Docs] Example of the wound area estimation
+
+### Update
+
+- :computer: [Python] Add new version of the semantic segmentation for the identification of the reference marker
+- :closed_book: [Docs] Update the old notebooks according to the new semantic segmentation model
+- :globe_with_meridians: [Global] Update online model checkpoint
+
+------------------------------------------------------------------------------
+
 ## [0.0.1] - 2023-11-08
 
 First version of the library.